+++ NEWS.md
@@ -14,11 +14,16 @@ Licence: see LICENCE.md
 
 - peak fitting (in particular for maps): working on a Scikit-Learn API
 
-# 0.6.2 (stable)
+# 0.6.3 (stable)
+
+  - fix 'roi' and 'method' arguments in baseline() function (issue #38)
+
+
+# 0.6.2
 
   - Fix the documentation.
 
-# 0.6.0 (stable)
+# 0.6.0
 
 ## Additions